+++ src/gl/utils/glQcmCatalogPanel.js
@@ -1,4 +1,0 @@
-export {
-  buildExportQuery,
-  buildQuestionsListQuery,
-} from '../../shared/qcm/qcmCatalogPanelQuery.js';